requirements.txt
# Core dependencies
python-telegram-bot==20.8
python-dotenv==1.0.0
openai==1.12.0

# HTTP client for async requests
aiohttp==3.9.3

# Optional dependencies for better performance
ujson==5.9.0
uvloop==0.19.0; sys_platform != 'win32'

# Development dependencies
pytest==8.0.2
pytest-asyncio==0.23.5
python-dotenv[cli]==1.0.0 